examples/benchmark-ui/requirements.txt
deepsparse-nightly
deepsparse[onnxruntime]
gradio==3.19.1
python-digitalocean>=1.17.0